+++ src/App/Controller/Logout.php
@@ -2,13 +2,10 @@
 namespace App\Controller;
 
 use Tk\Request;
-use Dom\Template;
 use Tk\Auth\AuthEvents;
 use Tk\Event\AuthEvent;
 
 /**
- * Class Index
- *
  * @author Michael Mifsud <info@tropotek.com>
  * @link http://www.tropotek.com/
  * @license Copyright 2015 Michael Mifsud
@@ -17,9 +14,7 @@ class Logout extends Iface
 {
 
     /**
-     *
      * @param Request $request
-     * @return Template
      */
     public function doDefault(Request $request)
     {
@@ -29,15 +24,4 @@ class Logout extends Iface
         \Tk\Uri::create('/')->redirect();
     }
 
-    /**
-     * Execute the renderer.
-     *
-     * @return mixed
-     */
-    public function show()
-    {
-        $template = $this->getTemplate();
-
-        return $this->getPage()->setPageContent($template);
-    }
 }