+++ src/App/Controller/StaticPage.php
@@ -2,9 +2,8 @@
 namespace App\Controller;
 
 use Tk\Request;
+
 /**
- * Class Index
- *
  * @author Michael Mifsud <info@tropotek.com>
  * @link http://www.tropotek.com/
  * @license Copyright 2015 Michael Mifsud
@@ -12,13 +11,6 @@ use Tk\Request;
 class StaticPage extends Iface
 {
 
-    /**
-     *
-     */
-    public function __construct()
-    {
-        parent::__construct();
-    }
 
     /**
      * @param Request $request
@@ -26,7 +18,9 @@ class StaticPage extends Iface
      */
     public function doDefault(Request $request)
     {
-        $page = new \App\Page\Page($this, $request->getAttribute('_staticPath'));
+        //$page = new \App\Page\Page($this, $request->getAttribute('_staticPath'));
+        vd($request->getAttributes());
+        $page = new \App\Page\Page();
         
         return $page;
     }